Added parallel substitutions
Instead of substituting single terms at a time, perform parallel substitutions
of the whole context. Normally, we have contexts which incrementally check
consequent terms, e.g.
[·, x:A, y:B, z:C]
means
[· ⊢ A], [·, x:A ⊢ y:B] and [·, x:A, y:B ⊢ z:C]
Instead, we create a new context, [Δ], such that every type in delta is well
formed under [Γ]. This simplifies the problem of substitution significantly,
particularly when we have terms typed under different contexts, such as types
which introduce binders (e.g. applying a function of type [Γ ⊢ Π(x : A). B] to
an argument [Γ ⊢ e : A] is problematic because [Γ ⊬ B[x/e]] in general)