Added Quotient types
Added quotient type constructors, projections into quotients and elimination
principles. Constructing a quotient type requires providing an underlying type
and an equivalence relation upon it. Elimination requires a witness that the
equivalence relation is preserved over a mapping-out principle.

Added a test for quotient types by constructing the relation [R ⊆ ℕ × ℕ] where
[R(0, 0)] and [∀n, m. R(S n, S m)] then using the proof that [R] is an
equivalence to take the quotient [ℕ / R] as a definition of the booleans. Then,
implemented a dependent [if] function using the quotient eliminator.