Changed structure of semantic values
Instead of a single data type representing semantic values, we now have a
separate data type for eliminators. Neutrals are now stored as a variable head
and a spine of eliminators. This allows for constant time lookup of the "reason"
of the block (which at the moment is always a neutral variable, but with pattern
unification may also be a metavariable)