Added pattern unification
Term holes can now be present in the syntax which will be elaborated into
metavariables solved during conversion checking. This allows a form of type
inference in the surface syntax. Added a few witness constructors to the surface
syntax representing equality proof transformations (symmetry, transitivity and
congruence) to assist in writing proofs.

Evaluation is now monadic as depends on a metacontext