Implemented definitional conversion rule for castrefl on convertible types.
Worth noting this rule is problematic because it only allows computation of
definitional equality of the form [cast(A, A, e, t) ≡ u] (or symmetric) which is
problematic when [u] is also a cast term -- how do we know whether to convert by
checking the types in the LHS are equal and then converting on arguments, or to
check if the casts convert between definitionally equal types? I believe the
first case subsumes the latter, but requires some fallback if conversion fails
which is annoying to implement.