Added observational equality between inductive types
This is a bit of a hack: we first create a unique tag for each syntactic
inductive type (plus unique tags for the "muted" inductive types used in Mendler
induction) and then say inductive types [I a ~ I' a'] compute to [a ~ a'] iff
their tags are equal, otherwise compute to the empty type. Cast now works
because there is enough structure in inductive types to determine at what point
to compose proofs of equality between the arguments.