Optimised semantic propositions
Previously, the whole environment was embedded into the semantic proposition
domain every time a propositional element was created. This happened despite the
environment being essentially the same every time, and had a massive performance
hit. Now we instead deal with parallel values with a semantic value and its
proposition, so a propositional environment is given by projections now.